Extract bitmap tests
